Dashing fixed, now in cancellable, vsyin not used

+++ Dev_class5_Orthogonal/Motor2D/SDL/include/SDL_rect.h
@@ -64,7 +64,7 @@ typedef struct SDL_Point
 typedef struct SDL_Rect
 {
     int x, y;
-    float w, h;
+    int w, h;
 } SDL_Rect;
 
 /**
